dashboard doctor and patient route create and design
